banco multiple 100%

+++ ajax/all_trabajador.php
@@ -21,6 +21,9 @@
 
       date_default_timezone_set('America/Lima');
       $date_now = date("d-m-Y h.i.s A");
+
+      $imagen_error = "this.src='../dist/svg/user_default.svg'";
+      $toltip = '<script> $(function () { $(\'[data-toggle="tooltip"]\').tooltip(); }); </script>';
       
       $idtrabajador	  	= isset($_POST["idtrabajador"])? limpiarCadena($_POST["idtrabajador"]):"";
       $nombre 		      = isset($_POST["nombre"])? limpiarCadena($_POST["nombre"]):"";
@@ -31,11 +34,12 @@
       $nacimiento		    = isset($_POST["nacimiento"])? limpiarCadena($_POST["nacimiento"]):"";
       $edad		          = isset($_POST["edad"])? limpiarCadena($_POST["edad"]):"";      
       $email			      = isset($_POST["email"])? limpiarCadena($_POST["email"]):"";
-      $banco			      = isset($_POST["banco"])? limpiarCadena($_POST["banco"]):"";      
-      $c_bancaria		    = isset($_POST["c_bancaria"])? limpiarCadena($_POST["c_bancaria"]):"";
-      $c_bancaria_format= isset($_POST["c_bancaria"])? limpiarCadena($_POST["c_bancaria"]):"";
-      $cci	          	= isset($_POST["cci"])? limpiarCadena($_POST["cci"]):"";
-      $cci_format      	= isset($_POST["cci"])? limpiarCadena($_POST["cci"]):"";
+      $banco_seleccionado= isset($_POST["banco_seleccionado"])? $_POST["banco_seleccionado"] :"";
+      $banco			      = isset($_POST["banco_array"])?$_POST["banco_array"]:"";      
+      $cta_bancaria		  = isset($_POST["cta_bancaria"])?$_POST["cta_bancaria"]:"";
+      $cta_bancaria_format= isset($_POST["cta_bancaria"])?$_POST["cta_bancaria"]:"";
+      $cci	          	= isset($_POST["cci"])?$_POST["cci"]:"";
+      $cci_format      	= isset($_POST["cci"])? $_POST["cci"]:"";
       $titular_cuenta		= isset($_POST["titular_cuenta"])? limpiarCadena($_POST["titular_cuenta"]):"";
       $tipo	          	= isset($_POST["tipo"])? limpiarCadena($_POST["tipo"]):"";
       $ocupacion	      = isset($_POST["ocupacion"])? limpiarCadena($_POST["ocupacion"]):"";
@@ -45,8 +49,8 @@
       $imagen2			    = isset($_POST["foto2"])? limpiarCadena($_POST["foto2"]):"";
       $imagen3			    = isset($_POST["foto3"])? limpiarCadena($_POST["foto3"]):"";
 
-      $cv_documentado			    = isset($_POST["doc4"])? limpiarCadena($_POST["doc4"]):"";
-      $cv_nodocumentado			  = isset($_POST["doc5"])? limpiarCadena($_POST["doc5"]):"";
+      $cv_documentado		= isset($_POST["doc4"])? limpiarCadena($_POST["doc4"]):"";
+      $cv_nodocumentado = isset($_POST["doc5"])? limpiarCadena($_POST["doc5"]):"";
       switch ($_GET["op"]) {
 
         case 'guardaryeditar':
@@ -128,7 +132,7 @@
 
           if (empty($idtrabajador)){
             
-            $rspta=$trabajador->insertar($nombre, $tipo_documento, $num_documento, $direccion, $telefono, $nacimiento, $edad,  $email, $banco, str_replace("-", "",$c_bancaria), $c_bancaria_format, str_replace("-", "",$cci), $cci_format, $titular_cuenta, $tipo, $ocupacion, $ruc, $imagen1, $imagen2, $imagen3, $cv_documentado, $cv_nodocumentado);
+            $rspta=$trabajador->insertar($nombre, $tipo_documento, $num_documento, $direccion, $telefono, $nacimiento, $edad,  $email, $banco_seleccionado, $banco, $cta_bancaria, $cci, $titular_cuenta, $tipo, $ocupacion, $ruc, $imagen1, $imagen2, $imagen3, $cv_documentado, $cv_nodocumentado);
             
             echo json_encode($rspta, true);
   
@@ -136,70 +140,40 @@
 
             // validamos si existe LA IMG para eliminarlo
             if ($flat_img1 == true) {
-
               $datos_f1 = $trabajador->obtenerImg($idtrabajador);
-
               $img1_ant = $datos_f1['data']['imagen_perfil'];
-
-              if ($img1_ant != "") {
-
-                unlink("../dist/docs/all_trabajador/perfil/" . $img1_ant);
-              }
+              if ($img1_ant != "") { unlink("../dist/docs/all_trabajador/perfil/" . $img1_ant);  }
             }
 
             //imagen_dni_anverso
             if ($flat_img2 == true) {
-
               $datos_f2 = $trabajador->obtenerImg($idtrabajador);
-
               $img2_ant = $datos_f2['data']['imagen_dni_anverso'];
-
-              if ($img2_ant != "") {
-
-                unlink("../dist/docs/all_trabajador/dni_anverso/" . $img2_ant);
-              }
+              if ($img2_ant != "") { unlink("../dist/docs/all_trabajador/dni_anverso/" . $img2_ant); }
             }
 
             //imagen_dni_reverso
             if ($flat_img3 == true) {
-
               $datos_f3 = $trabajador->obtenerImg($idtrabajador);
-
               $img3_ant = $datos_f3['data']['imagen_dni_reverso'];
-
-              if ($img3_ant != "") {
-
-                unlink("../dist/docs/all_trabajador/dni_reverso/" . $img3_ant);
-              }
+              if ($img3_ant != "") { unlink("../dist/docs/all_trabajador/dni_reverso/" . $img3_ant); }
             }
 
             //cvs
             if ($flat_cv1 == true) {
-
               $datos_cv1 = $trabajador->obtenercv($idtrabajador);
-
               $cv1_ant = $datos_cv1['data']['cv_documentado'];
-
-              if ($cv1_ant != "") {
-
-                unlink("../dist/docs/all_trabajador/cv_documentado/" . $cv1_ant);
-              }
+              if ($cv1_ant != "") { unlink("../dist/docs/all_trabajador/cv_documentado/" . $cv1_ant); }
             }
 
             if ($flat_cv2 == true) {
-
               $datos_cv2 = $trabajador->obtenercv($idtrabajador);
-
               $cv2_ant = $datos_cv2['data']['cv_no_documentado'];
-
-              if ($cv2_ant != "") {
-
-                unlink("../dist/docs/all_trabajador/cv_no_documentado/" . $cv2_ant);
-              }
+              if ($cv2_ant != "") { unlink("../dist/docs/all_trabajador/cv_no_documentado/" . $cv2_ant); }
             }
 
             // editamos un trabajador existente
-            $rspta=$trabajador->editar($idtrabajador, $nombre, $tipo_documento, $num_documento, $direccion, $telefono, $nacimiento, $edad, $email, $banco, str_replace("-", "",$c_bancaria), $c_bancaria_format, str_replace("-", "",$cci), $cci_format, $titular_cuenta, $tipo, $ocupacion, $ruc, $imagen1, $imagen2, $imagen3, $cv_documentado, $cv_nodocumentado);
+            $rspta=$trabajador->editar($idtrabajador, $nombre, $tipo_documento, $num_documento, $direccion, $telefono, $nacimiento, $edad, $email, $banco_seleccionado, $banco, $cta_bancaria, $cci,  $titular_cuenta, $tipo, $ocupacion, $ruc, $imagen1, $imagen2, $imagen3, $cv_documentado, $cv_nodocumentado);
             
             echo json_encode($rspta, true);
           }            
@@ -240,42 +214,40 @@
 
         case 'tbla_principal':          
 
-          $rspta=$trabajador->tbla_principal();
+          $rspta=$trabajador->tbla_principal(1);
+          
           //Vamos a declarar un array
           $data= Array(); $cont=1;
 
-          $imagen_error = "this.src='../dist/svg/user_default.svg'";
-          $toltip = '<script> $(function () { $(\'[data-toggle="tooltip"]\').tooltip(); }); </script>';
-          if ($rspta['status']) {
+          if ($rspta['status'] == true) {
 
             foreach ($rspta['data'] as $key => $value) {             
           
               $data[]=array(
                 "0"=>$cont++,
                 "1"=>($value['estado'])?'<button class="btn btn-warning btn-sm" onclick="mostrar('.$value['idtrabajador'].')" data-toggle="tooltip" data-original-title="Editar"><i class="fas fa-pencil-alt"></i></button>'.
-                  ' <button class="btn btn-danger btn-sm" onclick="eliminar('.$value['idtrabajador'].', \''.encodeCadenaHtml($value['nombres']).'\')" data-toggle="tooltip" data-original-title="Eliminar o papelera"><i class="fas fa-skull-crossbones"></i></button>'.
+                  ' <button class="btn btn-danger btn-sm" onclick="eliminar('.$value['idtrabajador'].', \''.encodeCadenaHtml($value['trabajador']).'\')" data-toggle="tooltip" data-original-title="Eliminar o papelera"><i class="fas fa-skull-crossbones"></i></button>'.
                   ' <button class="btn btn-info btn-sm" onclick="verdatos('.$value['idtrabajador'].')"data-toggle="tooltip" data-original-title="ver datos"><i class="far fa-eye"></i></button>':
                   ' <button class="btn btn-info btn-sm" onclick="verdatos('.$value['idtrabajador'].')" data-toggle="tooltip" data-original-title="Ver datos"><i class="far fa-eye"></i></button>',
                 "2"=>'<div class="user-block">
                   <img class="img-circle" src="../dist/docs/all_trabajador/perfil/'. $value['imagen_perfil'] .'" alt="User Image" onerror="'.$imagen_error.'">
-                  <span class="username"><p class="text-primary"style="margin-bottom: 0.2rem !important"; >'. $value['nombres'] .'</p></span>
+                  <span class="username"><p class="text-primary m-b-02rem" >'. $value['trabajador'] .'</p></span>
                   <span class="description">'. $value['tipo_documento'] .': '. $value['numero_documento'] .' </span>
                   </div>',
                 "3"=> $value['nombre_tipo'],
                 "4"=> $value['nombre_ocupacion'],
                 "5"=>'<a href="tel:+51'.quitar_guion($value['telefono']).'" data-toggle="tooltip" data-original-title="Llamar al trabajador.">'. $value['telefono'] . '</a>',
-                "6"=>format_d_m_a($value['fecha_nacimiento']).'<b>: </b>'.calculaedad($value['fecha_nacimiento']),
-                "7"=> '<b>'.$value['banco'] .': </b>'. $value['cuenta_bancaria_format'] .' <br> <b>CCI: </b>'.$value['cci_format'],
-                "8"=>(($value['estado'])?'<span class="text-center badge badge-success">Activado</span>':
-                '<span class="text-center badge badge-danger">Desactivado</span>').$toltip,
-                "9"=> $value['nombres'],
+                "6"=>format_d_m_a($value['fecha_nacimiento']).'<b>: </b>'. '<b>'.calculaedad($value['fecha_nacimiento']).'</b>' ,
+                "7"=> '<b>'.$value['banco'] .': </b>'. $value['cuenta_bancaria'] .' <br> <b>CCI: </b>'.$value['cci'],
+                "8"=>(($value['estado'])?'<span class="text-center badge badge-success">Activado</span>': '<span class="text-center badge badge-danger">Desactivado</span>').$toltip,
+                "9"=> $value['trabajador'],
                 "10"=> $value['tipo_documento'],
                 "11"=> $value['numero_documento'],
                 "12"=> format_d_m_a($value['fecha_nacimiento']),
                 "13"=>calculaedad($value['fecha_nacimiento']),
                 "14"=> $value['banco'],
-                "15"=> $value['cuenta_bancaria_format'],
-                "16"=> $value['cci_format'],
+                "15"=> $value['cuenta_bancaria'],
+                "16"=> $value['cci'],
               );
             }
             $results = array(
@@ -283,21 +255,18 @@
               "iTotalRecords"=>count($data), //enviamos el total registros al datatable
               "iTotalDisplayRecords"=>1, //enviamos el total registros a visualizar
               "data"=>$data);
-            echo json_encode($results);
+            echo json_encode($results, true);
 
           } else {
             echo $rspta['code_error'] .' - '. $rspta['message'] .' '. $rspta['data'];
           }
-
         break;  
 
         case 'listar_expulsado':
-          $rspta=$trabajador->listar_expulsado();
+          $rspta=$trabajador->tbla_principal(0);
           //Vamos a declarar un array
           $data= Array(); $cont=1;
-
-          $imagen_error = "this.src='../dist/svg/user_default.svg'";
-          $toltip = '<script> $(function () { $(\'[data-toggle="tooltip"]\').tooltip(); }); </script>';
+          
           if ($rspta['status']) {
 
             foreach ($rspta['data'] as $key => $value) { 
@@ -307,11 +276,11 @@
                   ' <button class="btn btn-danger btn-sm" onclick="desactivar('.$value['idtrabajador'].')" data-toggle="tooltip" data-original-title="Eliminar o papelera"><i class="far fa-trash-alt  "></i></button>'.
                   ' <button class="btn btn-info btn-sm" onclick="verdatos('.$value['idtrabajador'].')" data-toggle="tooltip" data-original-title="Ver datos"><i class="far fa-eye"></i></button>':
                   '<button class="btn btn-warning btn-sm" onclick="mostrar('.$value['idtrabajador'].')" data-toggle="tooltip" data-original-title="Editar"><i class="fa fa-pencil-alt"></i></button>'.
-                  ' <button class="btn btn-primary btn-sm" onclick="activar('.$value['idtrabajador'].', \''.encodeCadenaHtml($value['nombres']).'\')" data-toggle="tooltip" data-original-title="Recuperar"><i class="fa fa-check"></i></button>'.
+                  ' <button class="btn btn-primary btn-sm" onclick="activar('.$value['idtrabajador'].', \''.encodeCadenaHtml($value['trabajador']).'\')" data-toggle="tooltip" data-original-title="Recuperar"><i class="fa fa-check"></i></button>'.
                   ' <button class="btn btn-info btn-sm" onclick="verdatos('.$value['idtrabajador'].')" data-toggle="tooltip" data-original-title="Mostrar"><i class="far fa-eye"></i></button>',
                 "2"=>'<div class="user-block">
                   <img class="img-circle" src="../dist/docs/all_trabajador/perfil/'. $value['imagen_perfil'] .'" alt="User Image" onerror="'.$imagen_error.'">
-                  <span class="username"><p class="text-primary m-b-02rem" >'. $value['nombres'] .'</p></span>
+                  <span class="username"><p class="text-primary m-b-02rem" >'. $value['trabajador'] .'</p></span>
                   <span class="description">'. $value['tipo_documento'] .': '. $value['numero_documento'] .'<br>'.format_d_m_a($value['fecha_nacimiento']).' : '.$value['edad'].' años</span>
                   </div>',
                 "3"=> '<div class="center-vertical">'. $value['nombre_tipo'] .'</div>',
@@ -320,14 +289,14 @@
                 "6"=> $value['descripcion_expulsion'] ,
                 "7"=>(($value['estado'])?'<span class="text-center badge badge-success">Activado</span>':
                 '<span class="text-center badge badge-danger">Desactivado</span>').$toltip,
-                "8"=> $value['nombres'],
+                "8"=> $value['trabajador'],
                 "9"=> $value['tipo_documento'],
                 "10"=> $value['numero_documento'],
                 "11"=> format_d_m_a($value['fecha_nacimiento']),
                 "12"=>calculaedad($value['fecha_nacimiento']),
                 "13"=> $value['banco'],
-                "14"=> $value['cuenta_bancaria_format'],
-                "15"=> $value['cci_format'],
+                "14"=> $value['cuenta_bancaria'],
+                "15"=> $value['cci'],
               );
             }
 
@@ -336,7 +305,7 @@
               "iTotalRecords"=>count($data), //enviamos el total registros al datatable
               "iTotalDisplayRecords"=>1, //enviamos el total registros a visualizar
               "data"=>$data);
-            echo json_encode($results);
+            echo json_encode($results, true);
 
           } else {
             echo $rspta['code_error'] .' - '. $rspta['message'] .' '. $rspta['data'];
@@ -350,12 +319,17 @@
           echo json_encode($rspta, true);
         break;        
 
-        case 'formato_banco':
-           
+        case 'formato_banco':           
           $rspta=$trabajador->formato_banco($_POST["idbanco"]);
           //Codificar el resultado utilizando json
-          echo json_encode($rspta, true);
-           
+          echo json_encode($rspta, true);           
+        break;
+
+        /* =========================== S E C C I O N   R E C U P E R A R   B A N C O S =========================== */
+        case 'recuperar_banco':           
+          $rspta=$trabajador->recuperar_banco();
+          //Codificar el resultado utilizando json
+          echo json_encode($rspta, true);           
         break;
       }
